Refactor auth middlewares
